+++ WebRankingProviderRepository/IRepository.cs
@@ -7,8 +7,8 @@ namespace WebRankingProviderRepository
     {
         IQueryable<T> List();
         bool Create(T item);
-        bool Delete(int id);
-        T Get(int id);
+        bool Delete(long id);
+        T Get(long id);
         bool SaveChanges();
     }
 }
